compile-time enforcement of printf strings

+++ firmware/console/binary/tunerstudio.cpp
@@ -188,7 +188,7 @@ void TunerStudio::handleCrc32Check(TsChannelBase *tsChannel, uint16_t offset, ui
 	const uint8_t* start = getWorkingPageAddr() + offset;
 
 	uint32_t crc = crc32(start, count);
-	efiPrintf("TS <- Get CRC offset %d count %d result %08x", offset, count, crc);
+	efiPrintf("TS <- Get CRC offset %d count %d result %08x", offset, count, (unsigned int)crc);
 
 	crc = SWAP_UINT32(crc);
 	tsChannel->copyAndWriteSmallCrcPacket((const uint8_t *) &crc, sizeof(crc));
@@ -464,7 +464,7 @@ static int tsProcessOne(TsChannelBase* tsChannel) {
 		/* send error only if previously we were in sync */
 		if (tsChannel->in_sync) {
 			efiPrintf("TunerStudio: command %c actual CRC %x/expected %x", tsChannel->scratchBuffer[0],
-					actualCrc, expectedCrc);
+					(unsigned int)actualCrc, (unsigned int)expectedCrc);
 			tunerStudioError(tsChannel, "ERROR: CRC issue");
 			sendErrorCode(tsChannel, TS_RESPONSE_CRC_FAILURE);
 			tsChannel->in_sync = false;
