+++ firmware/controllers/lua/can_filter.cpp
@@ -31,7 +31,7 @@ void addLuaCanRxFilter(int32_t eid, uint32_t mask, CanBusIndex bus, int callback
 		firmwareError(ObdCode::OBD_PCM_Processor_Fault, "Too many Lua CAN RX filters");
 	}
 
-	efiPrintf("Added Lua CAN RX filter id 0x%x mask 0x%x with%s custom function", eid, mask, (callback == -1 ? "out" : ""));
+	efiPrintf("Added Lua CAN RX filter id 0x%x mask 0x%x with%s custom function", (unsigned int)eid, (unsigned int)mask, (callback == -1 ? "out" : ""));
 
 	filters[filterCount].Id = eid;
 	filters[filterCount].Mask = mask;
