+++ firmware/hw_layer/drivers/can/can_msg_tx.cpp
@@ -66,9 +66,9 @@ CanTxMessage::~CanTxMessage() {
 		efiPrintf("Sending CAN bus%d message: ID=%x/l=%x %x %x %x %x %x %x %x %x",
 				busIndex,
 #ifndef STM32H7XX
-				(m_frame.IDE == CAN_IDE_EXT) ? CAN_EID(m_frame) : CAN_SID(m_frame),
+				(unsigned int)((m_frame.IDE == CAN_IDE_EXT) ? CAN_EID(m_frame) : CAN_SID(m_frame)),
 #else
-						  m_frame.common.XTD ? CAN_EID(m_frame) : CAN_SID(m_frame),
+				(unsigned int)(m_frame.common.XTD ? CAN_EID(m_frame) : CAN_SID(m_frame)),
 #endif
 				m_frame.DLC,
 				m_frame.data8[0], m_frame.data8[1],
